Feat: implemented decals

+++ Doom/math.c
@@ -1,6 +1,6 @@
 #include "math.h"
 
-i32 get_line_intersection(v2 p0, v2 p1, v2 p2, v2 p3, v2* i) {
+u8 get_line_intersection(v2 p0, v2 p1, v2 p2, v2 p3, v2* i) {
 	v2 s1 = { p1.x - p0.x, p1.y - p0.y };
 	v2 s2 = { p3.x - p2.x, p3.y - p2.y };
