+++ Doom/math.h
@@ -10,7 +10,7 @@
 #define BOXINTERSECT2D(x0, y0, x1, y1, x2, y2, x3, y3) ((OVERLAP1D(x0, x1, x2, x3) && OVERLAP1D(y0, y1, y2, y3))
 #define POINTSIDE2D(px, py, x0, y0, x1, y1) ((x1 - x0) * (py - y0) - (y1 - y0) * (px - x0)) 
 
-i32 get_line_intersection(v2 p0, v2 p1, v2 p2, v2 p3, v2* i);
+u8 get_line_intersection(v2 p0, v2 p1, v2 p2, v2 p3, v2* i);
 v2 world_pos_to_camera(v2 pos, Player player);
 v2 camera_pos_to_world(v2 pos, Player player);
 i32 screen_angle_to_x(f32 angle);
